# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Folio API
  version: 0.0.0
  description: |
    Folio – personal finance backend.

    This file is the single source of truth for API contracts. Both the Go
    server (via oapi-codegen) and the TypeScript client (via openapi-typescript)
    are generated from it.
servers:
  - url: http://localhost:8080
    description: Local dev
  - url: https://api.folio.example.com
    description: Production
...
security:
  - sessionCookie: []
paths:
...
  /api/v1/onboarding/bootstrap:
    post:
      tags: [onboarding]
      summary: Bootstrap a tenant and its first user
      description: Creates the first tenant/user pair. Password auth is intentionally not implemented yet.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BootstrapInput" }
      responses:
        "201":
          description: Bootstrapped
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BootstrapResult" }
...
  /api/v1/me:
    get:
      tags: [auth]
      summary: Current tenant user
      security:
        - tenantHeader: []
      responses:
        "200":
          description: Current user and tenant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeResult" }
  /api/v1/accounts:
    get:
      tags: [accounts]
      summary: List accounts for current tenant
      security:
        - tenantHeader: []
      parameters:
        - name: includeArchived
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Account" }
    post:
      tags: [accounts]
      summary: Create an account and opening balance snapshot
      security:
        - tenantHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccountCreateInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
...
  /api/v1/transactions:
    get:
      tags: [transactions]
      summary: List transactions for current tenant
      description: |
        Tenant is taken from the `X-Tenant-ID` context — never from the body.
        Results are ordered by `bookedAt DESC, id DESC`.
      security:
        - tenantHeader: []
      parameters:
        - name: accountId
          in: query
          schema: { type: string, format: uuid }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/TransactionStatus" }
        - name: uncategorized
          in: query
          description: |
            When true, return only transactions with `category_id IS NULL`
            AND no `transaction_lines` for the transaction. Split
            transactions are considered categorized via their lines and are
            excluded.
          schema: { type: boolean, default: false }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Transaction" }
    post:
      tags: [transactions]
      summary: Create a manual transaction
      description: |
        Manual ledger entry. `tenantId` is resolved from `X-Tenant-ID`.
        `currency` must match the account currency (also enforced by a
        database trigger). Status defaults to `posted`.
      security:
        - tenantHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TransactionCreateInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Transaction" }
...
  /api/v1/categories:
    get:
      tags: [categories]
      summary: List categories for current tenant
      security:
        - tenantHeader: []
      parameters:
        - name: includeArchived
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Category" }
    post:
      tags: [categories]
      summary: Create a category
      security:
        - tenantHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategoryCreateInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
...
  /api/v1/merchants:
    get:
      tags: [merchants]
      summary: List merchants for current tenant
      security:
        - tenantHeader: []
      parameters:
        - name: includeArchived
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Merchants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Merchant" }
    post:
      tags: [merchants]
      summary: Create a merchant
      security:
        - tenantHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MerchantCreateInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Merchant" }
...
  /api/v1/tags:
    get:
      tags: [tags]
      summary: List tags for current tenant
      security:
        - tenantHeader: []
      parameters:
        - name: includeArchived
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Tags
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Tag" }
    post:
      tags: [tags]
      summary: Create a tag
      security:
        - tenantHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TagCreateInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tag" }
...
  /api/v1/categorization-rules:
    get:
      tags: [categorization-rules]
      summary: List categorization rules for current tenant
      description: |
        Rules are returned ordered by `priority` ascending, then `createdAt`
        ascending. All rules (enabled and disabled) are listed by default.
      security:
        - tenantHeader: []
      parameters:
        - name: enabled
          in: query
          description: |
            Optional filter. When `true`, only enabled rules are returned;
            when `false`, only disabled rules. Any other value is ignored.
          schema: { type: boolean }
      responses:
        "200":
          description: Rules
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CategorizationRule" }
    post:
      tags: [categorization-rules]
      summary: Create a categorization rule
      security:
        - tenantHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CategorizationRuleCreateInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategorizationRule" }
...
  /api/v1/transactions/{transactionId}/apply-categorization-rules:
    post:
      tags: [transactions, categorization-rules]
      summary: Apply the first matching categorization rule to a transaction
      description: |
        Evaluates enabled rules in `priority` ASC, then `createdAt` ASC order
        and applies the first match. Manual overrides win: `categoryId`,
        `merchantId`, and `countAsExpense` on the transaction are written
        only when they are currently null. Tags in `addTagIds` are always
        added idempotently. On match, the rule's `lastMatchedAt` is stamped
        with the current time. When no rule matches, the transaction is
        returned unchanged with `matched: false`.
      security:
        - tenantHeader: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Apply result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuleApplyResult" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: folio_session
    tenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: Temporary development tenant resolver until session auth lands.
  schemas:
...
    User:
      type: object
      required: [id, tenantId, email, displayName, createdAt]
      properties:
        id:          { type: string, format: uuid }
        tenantId:    { type: string, format: uuid }
        email:       { type: string, format: email }
        displayName: { type: string }
        createdAt:   { type: string, format: date-time }
    Tenant:
      type: object
      required: [id, name, baseCurrency, cycleAnchorDay, locale, timezone, createdAt]
      properties:
        id:             { type: string, format: uuid }
        name:           { type: string }
        baseCurrency:   { type: string, example: CHF }
        cycleAnchorDay: { type: integer, minimum: 1, maximum: 31 }
        locale:         { type: string, example: en-CH }
        timezone:       { type: string, example: Europe/Zurich }
        createdAt:      { type: string, format: date-time }
    BootstrapInput:
      type: object
      required: [tenantName, baseCurrency, locale, email, displayName]
      properties:
        tenantName:     { type: string, minLength: 1 }
        baseCurrency:   { type: string, example: CHF }
        cycleAnchorDay: { type: integer, minimum: 1, maximum: 31, default: 1 }
        locale:         { type: string, example: en-CH }
        timezone:       { type: string, example: Europe/Zurich, default: UTC }
        email:          { type: string, format: email }
        displayName:    { type: string, minLength: 1, maxLength: 80 }
    BootstrapResult:
      type: object
      required: [tenant, user]
      properties:
        tenant: { $ref: "#/components/schemas/Tenant" }
        user:   { $ref: "#/components/schemas/User" }
    MeResult:
      type: object
      required: [tenant, user]
      properties:
        tenant: { $ref: "#/components/schemas/Tenant" }
        user:   { $ref: "#/components/schemas/User" }
...
    AccountKind:
      type: string
      enum: [checking, savings, cash, credit_card, brokerage, crypto_wallet, loan, mortgage, asset, pillar_2, pillar_3a, other]
    Account:
      type: object
      required: [id, tenantId, name, kind, currency, openDate, openingBalance, openingBalanceDate, includeInNetworth, includeInSavingsRate, createdAt, updatedAt, balance]
      properties:
        id:                   { type: string, format: uuid }
        tenantId:             { type: string, format: uuid }
        name:                 { type: string }
        nickname:             { type: string, nullable: true }
        kind:                 { $ref: "#/components/schemas/AccountKind" }
        currency:             { type: string }
        institution:          { type: string, nullable: true }
        openDate:             { type: string, format: date-time }
        closeDate:            { type: string, format: date-time, nullable: true }
        openingBalance:       { type: string, description: "Decimal as string" }
        openingBalanceDate:   { type: string, format: date-time }
        includeInNetworth:    { type: boolean }
        includeInSavingsRate: { type: boolean }
        archivedAt:           { type: string, format: date-time, nullable: true }
        createdAt:            { type: string, format: date-time }
        updatedAt:            { type: string, format: date-time }
        balance:              { type: string, description: "Derived balance as decimal string: latest snapshot + posted/reconciled transactions booked on or after the snapshot's UTC date." }
        balanceAsOf:          { type: string, format: date-time, nullable: true, description: "as_of of the latest balance snapshot anchoring the derivation." }
    AccountCreateInput:
      type: object
      required: [name, kind, currency, openDate]
      properties:
        name:                 { type: string }
        nickname:             { type: string, nullable: true }
        kind:                 { $ref: "#/components/schemas/AccountKind" }
        currency:             { type: string }
        institution:          { type: string, nullable: true }
        openDate:             { type: string, format: date }
        openingBalance:       { type: string, default: "0" }
        openingBalanceDate:   { type: string, format: date }
        includeInNetworth:    { type: boolean }
        includeInSavingsRate: { type: boolean }
...
    TransactionStatus:
      type: string
      enum: [draft, posted, reconciled, voided]
    Transaction:
      type: object
      required: [id, tenantId, accountId, status, bookedAt, amount, currency, createdAt, updatedAt]
      properties:
        id:               { type: string, format: uuid }
        tenantId:         { type: string, format: uuid }
        accountId:        { type: string, format: uuid }
        status:           { $ref: "#/components/schemas/TransactionStatus" }
        bookedAt:         { type: string, format: date }
        valueAt:          { type: string, format: date, nullable: true }
        postedAt:         { type: string, format: date-time, nullable: true }
        amount:           { type: string, description: "Signed decimal as string, always in the account currency." }
        currency:         { type: string, description: "Must match the account currency." }
        originalAmount:   { type: string, nullable: true, description: "Pre-FX amount when booked in a foreign currency." }
        originalCurrency: { type: string, nullable: true }
        merchantId:       { type: string, format: uuid, nullable: true }
        categoryId:       { type: string, format: uuid, nullable: true }
        counterpartyRaw:  { type: string, nullable: true }
        description:      { type: string, nullable: true }
        notes:            { type: string, nullable: true }
        countAsExpense:   { type: boolean, nullable: true, description: "User override. null defers to the derived rule." }
        createdAt:        { type: string, format: date-time }
        updatedAt:        { type: string, format: date-time }
    TransactionCreateInput:
      type: object
      required: [accountId, bookedAt, amount, currency]
      properties:
        accountId:        { type: string, format: uuid }
        status:           { $ref: "#/components/schemas/TransactionStatus", default: posted }
        bookedAt:         { type: string, format: date }
        valueAt:          { type: string, format: date, nullable: true }
        postedAt:         { type: string, format: date-time, nullable: true }
        amount:           { type: string, description: "Signed decimal as string." }
        currency:         { type: string }
        categoryId:       { type: string, format: uuid, nullable: true }
        merchantId:       { type: string, format: uuid, nullable: true }
        counterpartyRaw:  { type: string, nullable: true }
        description:      { type: string, nullable: true }
        notes:            { type: string, nullable: true }
        countAsExpense:   { type: boolean, nullable: true }
...
    Category:
      type: object
      required: [id, tenantId, name, sortOrder, createdAt, updatedAt]
      properties:
        id:         { type: string, format: uuid }
        tenantId:   { type: string, format: uuid }
        parentId:   { type: string, format: uuid, nullable: true }
        name:       { type: string }
        color:      { type: string, nullable: true }
        sortOrder:  { type: integer }
        archivedAt: { type: string, format: date-time, nullable: true }
        createdAt:  { type: string, format: date-time }
        updatedAt:  { type: string, format: date-time }
    CategoryCreateInput:
      type: object
      required: [name]
      properties:
        parentId:  { type: string, format: uuid, nullable: true }
        name:      { type: string, minLength: 1 }
        color:     { type: string, nullable: true }
        sortOrder: { type: integer, default: 0 }
...
    Merchant:
      type: object
      required: [id, tenantId, canonicalName, createdAt, updatedAt]
      properties:
        id:                { type: string, format: uuid }
        tenantId:          { type: string, format: uuid }
        canonicalName:     { type: string }
        logoUrl:           { type: string, nullable: true }
        defaultCategoryId: { type: string, format: uuid, nullable: true }
        industry:          { type: string, nullable: true }
        website:           { type: string, nullable: true }
        notes:             { type: string, nullable: true }
        archivedAt:        { type: string, format: date-time, nullable: true }
        createdAt:         { type: string, format: date-time }
        updatedAt:         { type: string, format: date-time }
    MerchantCreateInput:
      type: object
      required: [canonicalName]
      properties:
        canonicalName:     { type: string, minLength: 1 }
        logoUrl:           { type: string, nullable: true }
        defaultCategoryId: { type: string, format: uuid, nullable: true }
        industry:          { type: string, nullable: true }
        website:           { type: string, nullable: true }
        notes:             { type: string, nullable: true }
...
    Tag:
      type: object
      required: [id, tenantId, name, createdAt, updatedAt]
      properties:
        id:         { type: string, format: uuid }
        tenantId:   { type: string, format: uuid }
        name:       { type: string }
        color:      { type: string, nullable: true }
        archivedAt: { type: string, format: date-time, nullable: true }
        createdAt:  { type: string, format: date-time }
        updatedAt:  { type: string, format: date-time }
    TagCreateInput:
      type: object
      required: [name]
      properties:
        name:  { type: string, minLength: 1 }
        color: { type: string, nullable: true }
...
    CategorizationRuleWhen:
      type: object
      description: |
        v1 match DSL. At least one field must be set. All provided fields are
        ANDed. Amount bounds are inclusive. `counterpartyContains` and
        `descriptionContains` are case-insensitive substring matches against
        the transaction's `counterpartyRaw` / `description`. `amountSign` is
        `"debit"` (amount < 0) or `"credit"` (amount > 0).
      properties:
        accountId:            { type: string, format: uuid }
        merchantId:           { type: string, format: uuid }
        counterpartyContains: { type: string, minLength: 1 }
        descriptionContains:  { type: string, minLength: 1 }
        amountMin:            { type: string, description: "Decimal as string (inclusive)." }
        amountMax:            { type: string, description: "Decimal as string (inclusive)." }
        amountSign:           { type: string, enum: [debit, credit] }
      additionalProperties: false
    CategorizationRuleThen:
      type: object
      description: |
        v1 action DSL. At least one field must be set. `countAsExpense`
        accepts `true`, `false`, or `null`. During apply, manual overrides
        win: category, merchant, and count-as-expense are only written when
        the transaction currently has no value; tags are added idempotently.
      properties:
        categoryId:     { type: string, format: uuid }
        merchantId:     { type: string, format: uuid }
        addTagIds:
          type: array
          items: { type: string, format: uuid }
          minItems: 1
        countAsExpense: { type: boolean, nullable: true }
      additionalProperties: false
    CategorizationRule:
      type: object
      required: [id, tenantId, priority, when, then, enabled, createdAt, updatedAt]
      properties:
        id:            { type: string, format: uuid }
        tenantId:      { type: string, format: uuid }
        priority:      { type: integer, description: "Lower numbers match first." }
        when:          { $ref: "#/components/schemas/CategorizationRuleWhen" }
        then:          { $ref: "#/components/schemas/CategorizationRuleThen" }
        enabled:       { type: boolean }
        lastMatchedAt: { type: string, format: date-time, nullable: true }
        createdAt:     { type: string, format: date-time }
        updatedAt:     { type: string, format: date-time }
    CategorizationRuleCreateInput:
      type: object
      required: [when, then]
      properties:
        priority: { type: integer, default: 1000 }
        enabled:  { type: boolean, default: true }
        when:     { $ref: "#/components/schemas/CategorizationRuleWhen" }
        then:     { $ref: "#/components/schemas/CategorizationRuleThen" }
...
    RuleApplyResult:
      type: object
      required: [matched]
      properties:
        matched: { type: boolean }
        ruleId:  { type: string, format: uuid, nullable: true, description: "Set when matched is true." }
        transaction:
          description: Post-apply transaction snapshot.
          $ref: "#/components/schemas/Transaction"